+++ src/main/java/org/generation/muebleria/repository/ImagenesProductoRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ImagenesProductoRepository extends JpaRepository<ImagenesProducto, Integer> {
+public interface ImagenesProductoRepository extends JpaRepository<ImagenesProducto, Long> {
 
     // Método para encontrar todas las imágenes por el ID del producto
-    List<ImagenesProducto> findByProductoIdProducto(Integer idProducto);
+    List<ImagenesProducto> findByProductoIdProducto(Long idProducto);
 
 }
